pyproject.toml
[tool.poetry]
name = "gao"
version = "0.1.0"
description = ""
authors = ["V. David Zvenyach <dave@zvenyach.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
lxml = "^4.9.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
vcrpy = "^4.2.1"
pytest-recording = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
